restruct the core.py, split it to gensql/ folder
