+++ MercurySQL/errors/orm/__init__.py
@@ -1,1 +1,1 @@
-import connpool_errors as connection_pool
+from . import connpool_errors as connection_pool
